+++ common/src/main/java/com/avp/common/block/resin/ResinVeinSpreaderConfig.java
@@ -4,13 +4,12 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.MultifaceSpreader;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.Fluids;
 import org.jetbrains.annotations.NotNull;
 
-import com.avp.common.block.AVPBlocks;
+import com.avp.common.entity.living.alien.AlienVariantTypes;
 
 class ResinVeinSpreaderConfig extends MultifaceSpreader.DefaultSpreaderConfig {
 
@@ -23,39 +22,55 @@ class ResinVeinSpreaderConfig extends MultifaceSpreader.DefaultSpreaderConfig {
 
     @Override
     public boolean stateCanBeReplaced(
-        BlockGetter blockGetter,
-        @NotNull BlockPos blockPos,
-        BlockPos blockPos2,
+        @NotNull BlockGetter blockGetter,
+        @NotNull BlockPos sourcePos,
+        @NotNull BlockPos destinationPos,
         @NotNull Direction direction,
-        @NotNull BlockState blockState
+        @NotNull BlockState destinationBlockState
     ) {
-        var blockState2 = blockGetter.getBlockState(blockPos2.relative(direction));
-        // TODO: Use tag here for resin nodes.
-        var isNotResinNode = !blockState2.is(AVPBlocks.IRRADIATED_RESIN_NODE.get()) && !blockState2.is(
-            AVPBlocks.ABERRANT_RESIN_NODE.get()
-        )
-            && !blockState2.is(AVPBlocks.NETHER_RESIN_NODE.get()) && !blockState2.is(AVPBlocks.RESIN_NODE.get());
-        // TODO: What about the other resin vein types?
-        var canReplace = !blockState2.is(AVPBlocks.RESIN.get()) && isNotResinNode && !blockState2.is(Blocks.MOVING_PISTON);
+        var alienVariantType = AlienVariantTypes.getForOrNull(block);
 
-        if (!canReplace) {
+        if (alienVariantType == null) {
+            // 'block' should always be a resin vein block, and therefore always have a variant type.
             return false;
         }
 
-        if (blockPos.distManhattan(blockPos2) == 2) {
-            var blockPos3 = blockPos.relative(direction.getOpposite());
+        // I don't normally leave commented-out code, but reverse-engineering this vein spreading was such a PITA
+        // that I don't really want to forget how to do this later on.
+        // var destinationSupportBlockPos = destinationPos.relative(direction);
+        // var destinationSupportBlockState = blockGetter.getBlockState(destinationSupportBlockPos);
+
+        // If the destination block state is replaceable for this variant (ex. enemy variant resin veins)...
+        var canDestinationBlockStateBeReplaced = destinationBlockState.is(alienVariantType.resinReplaceableTag())
+            // OR the state is air (in which case it's free real-estate)
+            || destinationBlockState.isAir()
+            // OR the state is a matching resin vein variant (allows same-space spreading, veins occupy multiple faces).
+            || (destinationBlockState.is(alienVariantType.resinVein().get()));
+
+        if (!canDestinationBlockStateBeReplaced) {
+            return false;
+        }
+
+        if (sourcePos.distManhattan(destinationPos) == 2) {
+            var blockPos3 = sourcePos.relative(direction.getOpposite());
 
             if (blockGetter.getBlockState(blockPos3).isFaceSturdy(blockGetter, blockPos3, direction)) {
                 return false;
             }
         }
 
-        var fluidState = blockState.getFluidState();
+        var fluidState = destinationBlockState.getFluidState();
 
-        if ((!fluidState.isEmpty() && !fluidState.is(Fluids.WATER)) || blockState.is(BlockTags.FIRE)) {
+        if ((!fluidState.isEmpty() && !fluidState.is(Fluids.WATER)) || destinationBlockState.is(BlockTags.FIRE)) {
             return false;
         } else {
-            return blockState.canBeReplaced() || super.stateCanBeReplaced(blockGetter, blockPos, blockPos2, direction, blockState);
+            return destinationBlockState.canBeReplaced() || super.stateCanBeReplaced(
+                blockGetter,
+                sourcePos,
+                destinationPos,
+                direction,
+                destinationBlockState
+            );
         }
     }
 
@@ -65,8 +80,8 @@ class ResinVeinSpreaderConfig extends MultifaceSpreader.DefaultSpreaderConfig {
     }
 
     @Override
-    public boolean isOtherBlockValidAsSource(BlockState blockState) {
-        // TODO: What about the other resin vein types?
-        return !blockState.is(AVPBlocks.RESIN_VEIN.get());
+    public boolean isOtherBlockValidAsSource(@NotNull BlockState blockState) {
+        var alienVariantType = AlienVariantTypes.getForOrNull(blockState);
+        return alienVariantType == null || !blockState.is(alienVariantType.resinVein().get());
     }
 }